CalculatorProject/requirements.txt
snet.sdk
grpcio
grpcio-tools
